# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello people
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /people/{id}:
    x-swagger-router-controller: people
    get:
      operationId: getOne
      description: get a person
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Sucess
          schema:
            $ref: "#/definitions/GetPersonResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update a person
      operationId: update
      parameters:
        - name: id
          description: Person id
          type: string
          in: path
          required: true
        - name: fname
          description: Person fname
          in: body
          required: true
          schema:
            $ref: "#/definitions/Person"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete: 
      description: delete a person
      operationId: delPerson
      parameters:
        - name: id
          description: Person id
          type: string
          in: path
          required:  true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  Person:
    type: object
    properties:
      fname:
        type: string
        description: task object fname
      lname:
        type: string
        description: task object lname
    required:
      - fname
      - lname
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  GetPersonResponse:
    required:
      - id
      - fname
      - lname
    properties:
      id:
        type: string
      fname:
        type: string
      lname:
        type: string